// packages/react/src/lib/components/footer-nav/footer-nav.d.ts
import * as React from 'react';
export declare const PdsFooterNav: React.FunctionComponent<Partial<{
    variant: "default" | "subtle";
    behavior: "3col" | "4col";
}> & {
    ref?: React.Ref<any> | undefined;
} & {
    style?: React.CSSProperties | undefined;
} & Partial<{}> & {
    id?: string | undefined;
    slot?: string | undefined;
    className?: string | undefined;
    children?: React.ReactNode;
}>;