// packages/react/src/lib/components/footer-nav-item/footer-nav-item.d.ts
import * as React from 'react';
export declare const PdsFooterNavItem: React.FunctionComponent<Partial<{
    label: string;
    variant: "default" | "subtle";
}> & {
    ref?: React.Ref<any> | undefined;
} & {
    style?: React.CSSProperties | undefined;
} & Partial<{}> & {
    id?: string | undefined;
    slot?: string | undefined;
    className?: string | undefined;
    children?: React.ReactNode;
}>;